requirements.txt
uvicorn==0.32.0
redis==5.1.1
python-dotenv==1.0.1
aiogram==3.13.1
pdfplumber==0.11.4
python-docx==1.1.2
redis==5.1.1
langchain==1.0.3
langchain-ollama==1.0.0
beautifulsoup4